reorganized graphics module
