+++ src/graphics/BaseGui.cpp
@@ -1,6 +1,6 @@
 #include <clashingcountries/graphics.hpp>
 
-using namespace clc_ui;
+using namespace clashing_countries::graphics;
 
 BaseGui::BaseGui() {
