+++ src/graphics/Container.cpp
@@ -1,5 +1,5 @@
 #include <clashingcountries/graphics.hpp>
-using namespace clc_ui;
+using namespace clashing_countries::graphics;
 
 void Container::draw(sf::RenderTarget &t, sf::RenderStates s) const {
